디렉터리 구성 변경(#8)
