+++ index.js
@@ -3,8 +3,7 @@
 //  */
 
 import {AppRegistry} from 'react-native';
-//import App from './App';
-import App from './src/components/Main';
+import App from './App';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
