+++ src/common/commom.text
@@ -1,0 +1,1 @@
+//공통된 컴포넌트를 위한 공간
