+++ src/components/login/Form.tsx
@@ -6,9 +6,9 @@ import {
   ImageBackground,
 } from 'react-native';
 import UserInput from './UserInput';
-import usernameImg from '../images/username.png';
-import passwordImg from '../images/password.png';
-import eyeImg from '../images/eye_black.png';
+import usernameImg from '../../../assets/images/username.png';
+import passwordImg from '../../../assets/images/password.png';
+import eyeImg from '../../../assets/images/eye_black.png';
 import {useState} from 'react';
 
 const Form = () => {
