+++ src/components/login/Wallpaper.tsx
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 // import Dimensions from 'Dimensions';
 import {StyleSheet, ImageBackground} from 'react-native';
 
-import bgSrc from '../images/wallpaper.png';
+import bgSrc from '../../../assets/images/wallpaper.png';
 
 export default class Wallpaper extends Component {
   render() {
