+++ src/constants/constants.text
@@ -1,0 +1,1 @@
+Folder to store any kind of constant that you have
