+++ src/helper/helper.text
@@ -1,0 +1,1 @@
+Common API를 위한 컨트롤러
