+++ src/redux/redux.text
@@ -1,0 +1,1 @@
+This folder should have all your reducers and store
